+++ finality-provider/service/chain_poller.go
@@ -115,7 +115,7 @@ func (cp *ChainPoller) blocksWithRetry(start, end uint64, limit uint32) ([]*type
 		// no need return error when just no found new blocks,
 		// the chain to poller may not produce new blocks.
 		if len(block) == 0 {
-			cp.logger.Warn(
+			cp.logger.Debug(
 				"no blocks found for range",
 				zap.Uint64("start_height", start),
 				zap.Uint64("end_height", end),
@@ -255,11 +255,19 @@ func (cp *ChainPoller) pollChain() {
 func (cp *ChainPoller) tryPollChain(latestBlockHeight, blockToRetrieve uint64) error {
 	var blocks []*types.BlockInfo
 	var err error
-	if blockToRetrieve == latestBlockHeight {
+
+	switch {
+	case blockToRetrieve > latestBlockHeight:
+		cp.logger.Debug(
+			"skipping block query as there is no new block",
+			zap.Uint64("next_height", blockToRetrieve),
+			zap.Uint64("latest_height", latestBlockHeight),
+		)
+	case blockToRetrieve == latestBlockHeight:
 		var latestBlock *types.BlockInfo
 		latestBlock, err = cp.consumerCon.QueryBlock(latestBlockHeight)
 		blocks = []*types.BlockInfo{latestBlock}
-	} else {
+	default:
 		blocks, err = cp.blocksWithRetry(blockToRetrieve, latestBlockHeight, cp.cfg.PollSize)
 	}